+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## Release candidate
 
+### Breaking changes
+- Dropped support for authorize query params `ldid` and `expiration` (#296, PLUM Sprint 231006)
+
 ### Features
 - Authorization for websocket requests (#300, PLUM Sprint 231006)
 - Silence stable log messages (#312, PLUM Sprint 231006)
 - External login registration webhook (#286, PLUM Sprint 231006)
+- OAuth Authorize ignores all unknown parameters (#296, PLUM Sprint 231006)
 
 ---